+++ core/setup/class-setup.php
@@ -121,11 +121,9 @@ if ( ! class_exists( '\WPOnion\Setup' ) ) {
 
 			if ( is_admin() && file_exists( WP_CONTENT_DIR . '/plugins/wponion/wponion.php' ) ) {
 				wponion_plugin_links( WPONION_FILE )
-					->action_link( 'docs', '<a href="https://docs.wponion.com">' . __( 'Documentation', 'wponion' ) . '</a>' )
-					->action_link( 'demo', __( 'Demo', 'wponion' ), 'https://wponion.com/demo' )
-					->row_link( __( 'Support', 'wponion' ), 'https://github.com/wponion' )
-					->row_link( __( 'Homepage', 'wponion' ), 'https://wponion.com' )
-					->row_link( __( 'Rate the plugin ★★★★★', 'wponion' ), 'https://wordpress.org/support/plugin/wponion/reviews/#new-post' );
+					->row_link( '<a href="https://docs.wponion.com">' . __( '📚 Documentation', 'wponion' ) . '</a>' )
+					->row_link( __( '🔍 Demo', 'wponion' ), 'https://wponion.com/demo' )
+					->row_link( __( '💬 Slack Community', 'wponion' ), 'https://wordpress.org/support/plugin/wponion/reviews/#new-post' );
 			}
 
 			self::register_core_fields();