+++ data/icons.php
@@ -1,5 +1,5 @@
 <?php
-/* Last Updated : 20/03/2020 - 08:23:53:am */
+/* Last Updated : 24/03/2020 - 07:21:10:am */
 
 use WPOnion\Helper as Helper;