+++ index.php
@@ -2,8 +2,8 @@
 /**
  * Plugin Name: WPOnion
  * Plugin URI: https://github.com/wponion/
- * Description: Powerfull,Reliable & Lightweight Options Framework.
- * Version: 1.4.1
+ * Description: ~ Lightweight, Flexible & Rapid WP Development Framework ~
+ * Version: 1.4.2
  * Author: varunsridharan
  * Author URI: https://varunsridharan.in
  * Text Domain: wponion